+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## Changelog
+### Version 0.0.6
+- Operations that have errors should no longer cause "Operation X already running"
+
 ### Version 0.0.5
 - Ability to set Bundle Id, Version Number, Build Number and Display Name of app
 - Handle when a user uses spaces in an app name when creating a project